🍎 MainActivity 코드 정리
## 🍎MainActivity 코드
- MainActivity 에서 관리하던 코드 별도의 `Composable` 을 만들어서 관리
